+++ WorldCities/WorldCitiesAPI/CHANGELOG.md
@@ -39,4 +39,6 @@ The following steps were used to generate this project:
 - Add `CountriesController.cs`
 - Modify `Program.cs` add JsonOptions to the AddController set JsonSerializerOptions.WeriteIndented to true
 - Modify `CitiesController.cs` GetCities method to implement a paging type functionality against the list of cities. 
-- Create new `Data/ApiResult.cs` file
+- Create new `Data/ApiResult.cs` file with pageIndex and pageSize
+- Modify `ApiResult.cs` add sorting functionality
+- Modify `CitiesController.cs` update GetCities with sorting functionality.